+++ GithubRepoFinder/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="11762" systemVersion="16C67" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="11762" systemVersion="16C67" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="mOq-FT-Wey">
     <device id="retina4_7" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
@@ -9,10 +9,26 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Repositories View Controller-->
+        <!--Navigation Controller-->
+        <scene sceneID="0fw-on-xOK">
+            <objects>
+                <navigationController id="mOq-FT-Wey" sceneMemberID="viewController">
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="JfS-FG-Jaq">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <connections>
+                        <segue destination="BYZ-38-t0r" kind="relationship" relationship="rootViewController" id="iwv-9K-Vik"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="pCt-9Q-8Me" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-761" y="139"/>
+        </scene>
+        <!--Swift Repos-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="RepositoriesViewController" customModule="GithubRepoFinder" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController title="Swift Repos" id="BYZ-38-t0r" customClass="RepositoriesViewController" customModule="GithubRepoFinder" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
@@ -50,13 +66,45 @@
                             <constraint firstItem="BRC-b4-GQ5" firstAttribute="leading" secondItem="8bC-Xf-vdC" secondAttribute="leading" id="hsV-36-uBV"/>
                         </constraints>
                     </view>
+                    <navigationItem key="navigationItem" title="Swift Repos" id="bxo-TX-dLJ"/>
                     <connections>
                         <outlet property="collectionView" destination="BRC-b4-GQ5" id="Je8-3y-X8C"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="136.80000000000001" y="138.98050974512745"/>
+            <point key="canvasLocation" x="73" y="139"/>
+        </scene>
+        <!--Pull Requests View Controller-->
+        <scene sceneID="rk2-st-k4r">
+            <objects>
+                <viewController id="0mA-TT-ZBx" customClass="PullRequestsViewController" customModule="GithubRepoFinder" customModuleProvider="target" sceneMemberID="viewController">
+                    <layoutGuides>
+                        <viewControllerLayoutGuide type="top" id="9FN-tn-NhF"/>
+                        <viewControllerLayoutGuide type="bottom" id="dK5-3h-YAL"/>
+                    </layoutGuides>
+                    <view key="view" contentMode="scaleToFill" id="6Ie-pj-DWV">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="WZI-nt-cQQ">
+                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                                <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                            </tableView>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="WZI-nt-cQQ" firstAttribute="leading" secondItem="6Ie-pj-DWV" secondAttribute="leading" id="AhF-dY-BA3"/>
+                            <constraint firstItem="WZI-nt-cQQ" firstAttribute="top" secondItem="6Ie-pj-DWV" secondAttribute="top" id="efD-NN-1QY"/>
+                            <constraint firstAttribute="trailing" secondItem="WZI-nt-cQQ" secondAttribute="trailing" id="j4K-vU-PQ0"/>
+                            <constraint firstItem="dK5-3h-YAL" firstAttribute="top" secondItem="WZI-nt-cQQ" secondAttribute="bottom" id="vcP-aO-7wl"/>
+                        </constraints>
+                    </view>
+                    <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" prompted="NO"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="Rj2-xa-HoU" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="842" y="139"/>
         </scene>
     </scenes>
 </document>
