+++ GithubRepoFinder/PullRequestCell.swift
@@ -1,0 +1,17 @@
+//
+//  PullRequestCell.swift
+//  GithubRepoFinder
+//
+//  Created by Jonathan Bijos on 19/01/17.
+//  Copyright © 2017 Quaggie. All rights reserved.
+//
+
+import UIKit
+
+class PullRequestCell: UITableViewCell {
+
+  override func awakeFromNib() {
+    super.awakeFromNib()
+    // Initialization code
+  }
+}
