+++ GithubRepoFinder/PullRequestsViewController.swift
@@ -1,0 +1,17 @@
+//
+//  PullRequestsViewController.swift
+//  GithubRepoFinder
+//
+//  Created by Jonathan Bijos on 19/01/17.
+//  Copyright © 2017 Quaggie. All rights reserved.
+//
+
+import UIKit
+
+class PullRequestsViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+  }
+
+}
