--- GithubRepoFinder/UIView+Helper.swift
@@ -1,9 +1,0 @@
-//
-//  UIView+Helper.swift
-//  GithubRepoFinder
-//
-//  Created by Jonathan Bijos on 19/01/17.
-//  Copyright © 2017 Quaggie. All rights reserved.
-//
-
-extension UIView: Gradientable {}
